rename util to vex_utility to avoid potential file name clashes
git-svn-id: https://svn.atnf.csiro.au/difx@7343 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876

+++ applications/vex2difx/trunk/vexdatamodel/vex_setup.cpp
@@ -159,9 +159,9 @@ void VexSetup::selectTones(enum ToneSelection selection, double guardBandMHz)
 	}
 }
 
-int VexSetup::nRecordChan() const
+size_t VexSetup::nRecordChan() const
 {
-	int rc = 0;
+	size_t rc = 0;
 
 	for(std::vector<VexStream>::const_iterator it = streams.begin(); it != streams.end(); ++it)
 	{
